# openapi.yaml
openapi: 3.0.3
info:
  title: CEC HTTP Bridge API
  description: |
    REST API for controlling HDMI-CEC devices (Rust implementation). Power,
    volume, source/HDMI switching, remote keys, raw CEC commands, MQTT
    settings, a dev console surface, and real-time events (SSE + WebSocket).
    Intended for home automation, media centers, and scripting.

    ## Envelope
    Every JSON response uses the envelope
    `{"status":"success"|"error","message":"...","data":...}`.
    `message` and `data` are omitted when empty/absent. Long-running
    operations return `success` with `data.accepted=true` (HTTP 202).

    ## Authentication
    Auth is **optional** and active only when the server was started with a
    token (`-token <secret>` / `auth_token` in `config.json`). When active,
    every `/api` route **except `GET /api/health`** (and `GET /metrics`)
    requires the token via `Authorization: Bearer <token>`,
    `X-Auth-Token: <token>`, `?key=<token>`, or the `capi_token` cookie set by
    the UI login page. Mutating cookie-authenticated requests additionally
    require same-origin. When no token is configured, no auth is required.
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local (default)
  - url: http://{host}:8080
    description: Network
    variables:
      host:
        default: localhost
security:
  - BearerAuth: []
...
paths:
...
  /api/dev/mode:
    get:
      tags: [Dev]
      summary: Current CEC mode
      description: >
        Returns `mode` ("passive" | "monitor_only"), `live_mode`,
        `activate_source`, and `adapter_connected`.
      operationId: getDevMode
      responses:
        '200':
          description: CEC mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Dev]
      summary: Set CEC mode
      description: >
        Body `{"mode": "passive" | "monitor_only"}`. Persists to config.json
        and signals the supervisor to reconnect so the mode takes effect.
      operationId: postDevMode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode:
                  type: string
                  enum: [passive, monitor_only]
      responses:
        '200':
          description: Mode updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/dev/probe:
    post:
      tags: [Dev]
      summary: Run Give* probes against a logical address
      description: |
        Runs the selected probes sequentially, each with its own observation
        window (default 600ms; tunable via `observe_ms`, hard-capped at
        5000ms) so replies are attributed to the probe that elicited them.
        `kind`: power | osd | vendor | cec_version | physical | audio | sam |
        menu | all. Rejected with 409 while the adapter is monitor-only.
      operationId: postDevProbe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address, kind]
              properties:
                address:
                  type: integer
                  minimum: 0
                  maximum: 15
                kind:
                  type: string
                  enum: [power, osd, vendor, cec_version, physical, audio, sam, menu, all]
                observe_ms:
                  type: integer
                  description: Per-probe observation window (default 600, clamped to <= 5000)
      responses:
        '200':
          description: Probe results (per-step replies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Adapter is in monitor-only mode
        '500':
          $ref: '#/components/responses/InternalError'
  /api/dev/send_key:
    post:
      tags: [Dev]
      summary: Low-level key send (always blind)
      description: |
        Transmits press/release even if the destination is not in the active
        mask. `hold_ms` clamped to <= 2000, `repeat` clamped to <= 32.
        Keycode 0 rejected; use `key: "select"`.
      operationId: postDevSendKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevSendKeyRequest'
      responses:
        '200':
          description: Send steps (press/release, acked flags)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Adapter is in monitor-only mode
        '500':
          $ref: '#/components/responses/InternalError'
  /api/dev/send_opcode:
    post:
      tags: [Dev]
      summary: Raw CEC opcode send (always blind)
      description: >
        Body `{"destination": 0, "opcode": 143, "params_hex": "01 02 ff"}`.
        Params accept spaced or compact hex. Rejected with 409 while the
        adapter is monitor-only.
      operationId: postDevSendOpcode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [destination, opcode]
              properties:
                destination:
                  type: integer
                  minimum: 0
                  maximum: 15
                opcode:
                  type: integer
                  minimum: 0
                  maximum: 255
                params_hex:
                  type: string
                  description: Hex byte string ("01 02 ff" or "0102ff")
      responses:
        '200':
          description: Opcode sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Adapter is in monitor-only mode
        '500':
          $ref: '#/components/responses/InternalError'
  /api/dev/run_strategies:
    post:
      tags: [Dev]
      summary: Run the strategy chain for an action
      description: |
        Body `{"action": "volume_up", "target": 5, "all_strategies": false,
        "observe_ms": 600}`. Runs are single-flight per connection: a second
        concurrent run is rejected instead of interleaving frames on the bus.
        `observe_ms` clamped to <= 5000.
      operationId: postDevRunStrategies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  example: volume_up
                target:
                  type: integer
                  minimum: 0
                  maximum: 15
                  nullable: true
                  description: null = follow the strategy's own targets
                all_strategies:
                  type: boolean
                observe_ms:
                  type: integer
      responses:
        '200':
          description: Strategy run results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Another strategy run is in flight, or adapter is monitor-only
        '500':
          $ref: '#/components/responses/InternalError'
  /api/dev/save_strategy:
    post:
      tags: [Dev]
      summary: Install a winning strategy per vendor
      description: >
        Body `{"vendor": "0x000048", "action": "volume_up",
        "strategy": "direct"}`. Persisted to config.json.
      operationId: postDevSaveStrategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vendor, action, strategy]
              properties:
                vendor:
                  type: string
                  example: "0x000048"
                action:
                  type: string
                  example: volume_up
                strategy:
                  type: string
                  description: Strategy name from the bench/run result
      responses:
        '200':
          description: Strategy saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/dev/actions:
    get:
      tags: [Dev]
      summary: List registry actions
      operationId: getDevActions
      responses:
        '200':
          description: Available actions and their strategies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/dev/keys:
    get:
      tags: [Dev]
      summary: List supported key names and keycodes
      operationId: getDevKeys
      responses:
        '200':
          description: Key name -> keycode table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/dev/opcodes:
    get:
      tags: [Dev]
      summary: List known CEC opcodes
      operationId: getDevOpcodes
      responses:
        '200':
          description: Opcode name/value table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  # ────────────────────────────────── Web UI ──────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Optional. Active only when the server runs with a token
        (`-token <secret>`). `X-Auth-Token`, `?key=`, and the `capi_token`
        cookie are accepted equivalently. When no token is configured this
        security requirement is a no-op.
...
  schemas:
    ApiResponse:
      type: object
      description: >
        Standard envelope. `message` omitted when empty; `data` omitted when
        absent. Async acceptance returns success with data.accepted=true.
      required: [status]
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
          description: Human-readable message (omitted when empty)
        data:
          description: Response payload (varies by endpoint; omitted when absent)
          nullable: true
...
    DevSendKeyRequest:
      type: object
      required: [address]
      properties:
        address:
          type: integer
          minimum: 0
          maximum: 14
          description: Destination (raw; never refused for "not active")
        key:
          type: string
          description: Canonical key name (preferred; see /api/dev/keys)
        keycode:
          type: integer
          minimum: 1
          maximum: 255
          description: Used when key is empty; 0 rejected
        hold_ms:
          type: integer
          description: Pause between press and release (0 = no release; clamped <= 2000)
        wait:
          type: boolean
          description: libcec wait-for-ack on press/release
        repeat:
          type: integer
          description: Press the key N times back-to-back (clamped <= 32)
  responses:
    BadRequest:
      description: Invalid parameters or request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: error
            message: Invalid logical address
    InternalError:
      description: CEC operation failed or server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: error
            message: failed to open adapter